Update models to better consistency
The Sequence Diagram was updated to show communication between the StaffLoginSystem, UserManagement, and OrganizerCreditialDatabase

The Domain Class Diagram was updated to show how the classes of the design should communicate. Multiplicities were updated and association names were changed to be  a better descriptor.

The System Class Model was changed to put classes in packages according to function of the program. Remote packages were labeled as remote systems.